09-pjt/BE/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.7.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==4.2.5
exceptiongroup==1.2.2
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
numpy==2.0.2
openai==1.61.1
Pillow==10.0.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
sniffio==1.3.1
sqlparse==0.4.4
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.3.0
Wikipedia-API==0.8.1
django-allauth==0.61.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
dj-rest-auth==7.0.0
PyJWT==2.9.0
requests==2.32.3